Make UnstructuredToLegacyFolder a method on FolderUnifiedStoreImpl
